Deepfacelab/requirements_clean.txt
absl-py==0.15.0
astunparse==1.6.3
attrs==24.2.0
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi 
charset-normalizer==3.3.2
click==8.1.8
cloudpickle==2.2.1
colorama==0.4.6
cycler==0.11.0
dask==2022.2.0
facenet-pytorch==2.5.3
ffmpeg-python==0.1.17
filelock==3.12.2
fire==0.7.0
Flask==2.2.5
Flask-Cors==5.0.0
flatbuffers==1.12
fonttools==4.38.0
fsspec==2023.1.0
future==1.0.0
gast==0.3.3
gdown==4.7.3
google-auth==2.29.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.32.0
gunicorn==23.0.0
h5py==2.10.0
idna==3.7
importlib-metadata==6.7.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.5
joblib==1.3.2
jsonschema==4.17.3
keras==2.11.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==18.1.1
locket==1.0.0
Markdown==3.4.4
MarkupSafe==2.1.5
matplotlib==3.5.3
mtcnn==0.1.1
networkx==2.6.3
numexpr==2.8.6
numpy==1.19.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
onnx==1.14.1
opencv-python==4.1.0.25
opt-einsum==3.3.0
packaging==24.0
pandas==1.3.5
partd==1.4.1
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
protobuf==3.20.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyWavelets==1.3.0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scikit-image==0.14.2
scikit-learn==1.0.2
scipy==1.4.1
six==1.15.0
soupsieve==2.4.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==1.1.0
tf2onnx==1.9.3
threadpoolctl==3.1.0
toolz==0.12.1
torch==1.13.1
torchvision==0.14.1
tqdm==4.66.2
typing-extensions==3.7.4.3
urllib3==2.0.7
Werkzeug==2.2.3
wrapt==1.12.1
zipp==3.15.0
